Create a interface to HTTP Verbs
To make all verbs have a single interface to access data (methods), the
creation of `VerbInterface` interface was required to declare only one
action to be taken, which can be implemented for each verb.

In `HTTP\Verb\Get` class has implemented the `action()` method and is
available for use in `Router`.