Serialize controller responses.
Controllers return a Zend\Http\Response object with a content, which can
be anything: a string, an object, an array, null.
Application creates the chain of responsability to handler the Accept
request serialization, NotAcceptable being the last handler on that
chain.

Controllers are required to provide a valid Response object, with the
proper HTTP status code and messsage as well as a content to be
serialized into the body.

This way, we have decoupled response creation from serialization. Being
able to re-use the current existing strategy for every controller which
will be created in the future.